refactor: 각 element들을 적절한 패키지로 이동
