+++ feature/main/src/main/java/com/practice/main/daily/components/DailyModeElements.kt
@@ -1,64 +1,30 @@
-package com.practice.main.daily
+package com.practice.main.daily.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.DatePicker
 import androidx.compose.material3.DatePickerFormatter
-import androidx.compose.material3.DatePickerState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.practice.designsystem.LightAndDarkPreview
-import com.practice.designsystem.LightPreview
 import com.practice.designsystem.components.BodySmall
 import com.practice.designsystem.theme.BlindarTheme
-import com.practice.main.daily.components.DailyDatePickerState
-import com.practice.main.daily.components.rememberDailyDatePickerState
-import com.practice.main.previewMainUiState
+import com.practice.main.daily.picker.DailyDatePickerState
+import com.practice.main.daily.picker.rememberDailyDatePickerState
 import kotlin.math.absoluteValue
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun DailyModeDatePicker(
-    datePickerState: DatePickerState,
-    onSelectDate: (Long) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    DisposableEffect(key1 = datePickerState.selectedDateMillis) {
-        onDispose {
-            datePickerState.selectedDateMillis?.let {
-                onSelectDate(it)
-            }
-        }
-    }
-
-    Column(modifier = modifier) {
-        DatePicker(
-            state = datePickerState,
-            dateFormatter = rememberDatePickerFormatter(),
-            showModeToggle = false,
-        )
-    }
-}
 
 @Composable
 internal fun DateQuickNavigationButtons(
@@ -119,27 +85,6 @@ private fun rememberDatePickerFormatter() = remember {
         selectedDateSkeleton = "yMMMdEEE",              // 2024년 1월 8일 (월)
         selectedDateDescriptionSkeleton = "yMMMdEEEE",  // 2024년 1월 8일 월요일
     )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@LightPreview
-@Composable
-private fun DailyModeDatePickerPreview() {
-    val datePickerState = rememberDailyModeDatePickerState(previewMainUiState())
-    var selectedMillis by remember { mutableLongStateOf(0L) }
-
-    BlindarTheme {
-        Column(modifier = Modifier.fillMaxWidth()) {
-            DailyModeDatePicker(
-                datePickerState = datePickerState,
-                onSelectDate = { selectedMillis = it },
-                modifier = Modifier.fillMaxWidth(),
-            )
-            Text(
-                text = selectedMillis.toString(),
-            )
-        }
-    }
 }
 
 @LightAndDarkPreview
